+++ framework/packages/abstract-core/src/error.rs
@@ -4,6 +4,8 @@ use cw_semver::Error as CwSemverError;
 use semver::{Error as SemverError, Version};
 use thiserror::Error;
 
+use crate::objects::{ans_host::AnsHostError, version_control::VersionControlError};
+
 /// Wrapper error for the Abstract framework.
 #[derive(Error, Debug, PartialEq)]
 pub enum AbstractError {
@@ -15,6 +17,12 @@ pub enum AbstractError {
 
     #[error("cw math overflow error: {0}")]
     Overflow(#[from] OverflowError),
+
+    #[error("{0}")]
+    VersionControlError(#[from] VersionControlError),
+
+    #[error("{0}")]
+    AnsHostError(#[from] AnsHostError),
 
     #[error("Semver error encountered while handling account object: {0}")]
     Semver(String),