+++ modules/contracts/apps/dca/tests/wasm.rs
@@ -1,5 +1,5 @@
 use dca_app::contract::DCA_APP_ID;
-use dca_app::DCAApp;
+use dca_app::DCA;
 
 use cw_orch::prelude::*;
 
@@ -14,7 +14,7 @@ fn successful_wasm() {
     let mock = Mock::new(&sender);
 
     // Construct the counter interface
-    let contract = DCAApp::new(DCA_APP_ID, mock);
+    let contract = DCA::new(DCA_APP_ID, mock);
 
     contract.wasm();
 }