+++ framework/CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `DepositMsgs` removed (now `deposit()` returns `Vec<CosmosMsg>`)
 - Abstract removed from the fields where it's redundant
 - InstantiateMsg is now removed from the install_adapter API
+- Removed `wasm_smart_query` helper, since it's accessible from `Querier` object
 
 ### Fixed